requirements.txt
google-cloud-speech
google-cloud-texttospeech
openai